Text_Evolution_with_Genetic_Algorithm
This project is a Python application that tries to find a target text sentence with evolutionary methods using a genetic algorithm. In this project, the aim is for a randomly initialized population to approach a certain target sentence through evolutionary processes. It is a project I did for the heuristic methods course.